pyproject.toml
[tool.poetry]
name = "parter-poker"
version = "0.1.0"
description = ""
authors = ["Lilith <__lilith@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
buttplug-py = "^0.2.0"
flask = "^3.0.3"
tqdm = "^4.66.4"
libsass = "^0.23.0"
waitress = "^3.0.0"
rapidfuzz = "^3.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = 'run.run:dev'
#dev = 'parter_poker.partner:devapp'
prod = 'run.run:prod'
#prod = 'waitress-serve --listen *:8080 --call parter_poker.partner:myapp'
#prod_ngrok = 'run.run:prod_serve'

[tool.poetry.plugins.dotenv]
ignore = "false"
location = ".env"